+++ Dnn.CommunityForums/class/ForumInfo.cs
@@ -18,22 +18,15 @@
 // DEALINGS IN THE SOFTWARE.
 //
 
-using System;
-using System.Data;
 
-using Microsoft.ApplicationBlocks.Data;
-namespace DotNetNuke.Modules.ActiveForums.Data
+using System;
+using System.Collections;
+using System.Collections.Generic;
+
+namespace DotNetNuke.Modules.ActiveForums
 {
-	[Obsolete("Deprecated in Community Forums. Removed in 10.00.00. Not Used.")]
-	public class Groups : DataConfig
-	{
-        //Public Function Forums_List(ByVal PortalId As Integer, ByVal ModuleId As Integer) As IDataReader
-        //    Return SqlHelper.ExecuteReader(_connectionString, dbPrefix & "Forums_GetPermissions", PortalId, ModuleId)
-        //End Function
-        [Obsolete("Deprecated in Community Forums. Removed in 10.00.00. Not Used.")]
-        public IDataReader Groups_Get(int ModuleId, int ForumGroupId)
-		{
-			return SqlHelper.ExecuteReader(_connectionString, dbPrefix + "Groups_Get", ModuleId, ForumGroupId);
-		}
-	}
+    [Obsolete("Deprecated in Community Forums. Removed in 10.00.00. Use DotNetNuke.Modules.ActiveForums.Entities.ForumInfo")]
+    public class Forum : DotNetNuke.Modules.ActiveForums.Entities.ForumInfo { }
+    [Obsolete("Deprecated in Community Forums. Removed in 10.00.00. Use DotNetNuke.Modules.ActiveForums.Entities.ForumInfo")]
+    public class ForumInfo : DotNetNuke.Modules.ActiveForums.Entities.ForumInfo { }
 }
