+++ Dnn.CommunityForums/components/Filters/FilterInfo.cs
